requirements.txt
fastapi
uvicorn
requests
elasticsearch
apache-airflow
pandas
python-dotenv
boto3
mysql-connector-python
pymongo